pyproject.toml
[project]
name = "nonebot-plugin-answersbook"
version = "0.2.1beta"
description = "The Book of Answers for NoneBot2"
authors = [
    {name = "Akirami", email = "Akiramiaya@outlook.com"},
]
license-expression = "MIT"
dependencies = ["nonebot2>=2.0.0b2", "nonebot-adapter-onebot>=2.0.0b1"]
requires-python = ">=3.8"
readme = "README.md"

[project.urls]
Homepage = "https://github.com/A-kirami/nonebot-plugin-answersbook"
Repository = "https://github.com/A-kirami/nonebot-plugin-answersbook"

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
